댓글 삭제 validation 추가

+++ src/app/controllers/commentController.js
@@ -37,8 +37,15 @@ exports.deleteComment = async function (req, res) {
     const commentId = req.params['commentId'];
     // 댓글 삭제시 댓글 목록 바로 업데이트해주는 트랜잭션필요
     try {
+        
         const userId = req.verifiedToken.id;
         const [userIdx] = await userDao.getUserIdxbyId(userId);
+
+        const [selectCommentRows] = await commentDao.selectComment(commentId);
+        if(!selectCommentRows) return res.json({isSucess : false, code : 320, message : "존재하지 않는 댓글입니다."});
+        const [selectCommentIsDeletedRows] = await commentDao.selectCommentIsDeleted(commentId);
+        if(!selectCommentIsDeletedRows) return res.json({isSucess : false, code : 320, message : "존재하지 않는 댓글입니다."});
+
         const deleteCommentParams = [commentId];
         const deleteCommmentRows = await commentDao.deleteComment(deleteCommentParams);
         return res.json({isSucess : true, code : 200, message : "댓글 삭제 성공"});
